feat: add lambda invoke ingest
